+++ src/libs/classNameCache.ts
@@ -2,15 +2,17 @@ import * as fs from "fs";
 import * as vscode from "vscode";
 import safeParser from "postcss-safe-parser";
 import selectorParser from "postcss-selector-parser";
-import Cache, { ClassNameData, ClassNameDataMap } from "./cache";
+import { DEBOUNCE_TIMER, SUPPORTED_MODULES } from "../config";
+import Cache from "./cache";
 import {
   getWorkspaceRelativeUriPath,
   resolveWorkspaceRelativePath,
 } from "../utils/getPath";
-import { DEBOUNCE_TIMER, SUPPORTED_MODULES } from "../config";
+import { sanitizeCssInput } from "../utils/sanitizeCssInput";
 import isPositionInComment from "../utils/isPositionInComment";
 import CssModuleDependencyCache from "./cssModuleDependencyCache";
 import CheckDocument from "./checkDocument";
+import { ClassNameRange, ClassNameRangeMap } from "../types/cache";
 
 /**
  * A utility class to extract and cache class names from CSS Module files.
@@ -136,7 +138,7 @@ export default class ClassNameCache {
   static async getClassNameDataFromImportPath(
     className: string,
     importPath: string
-  ): Promise<ClassNameData[] | undefined> {
+  ): Promise<ClassNameRange[] | undefined> {
     if (Cache.classNameCache.hasByKey(importPath)) {
       return Cache.classNameCache.getByKey(importPath)?.get(className);
     } else {
@@ -208,7 +210,7 @@ export default class ClassNameCache {
    */
   static async extractAndCacheClassNames(
     importPath: string
-  ): Promise<ClassNameDataMap | undefined> {
+  ): Promise<ClassNameRangeMap | undefined> {
     const filePath = resolveWorkspaceRelativePath(importPath);
     if (!fs.existsSync(filePath)) {
       Cache.classNameCache.deleteByKey(importPath); // Clean up stale entries
@@ -221,8 +223,9 @@ export default class ClassNameCache {
     }
 
     const content = document.getText();
-    const root = safeParser(content);
-    const classNames = new ClassNameDataMap();
+    const sanitizedContent = sanitizeCssInput(content);
+    const root = safeParser(sanitizedContent);
+    const classNames = new ClassNameRangeMap();
     const rules: Parameters<Parameters<(typeof root)["walkRules"]>[0]>[0][] =
       [];
 
@@ -255,7 +258,7 @@ export default class ClassNameCache {
               new vscode.Position(ruleStart.line - 1, ruleStart.column - 1)
             ) + sourceIndex;
 
-          const data: ClassNameData = {
+          const data: ClassNameRange = {
             range: new vscode.Range(
               document.positionAt(selectorOffsetInDoc),
               document.positionAt(selectorOffsetInDoc + value.length + 1) // +1 for "."